+++ CHANGES.md
@@ -8,6 +8,8 @@
 - #3829: Rich text from LibreOffice Calc is sent as screenshots
 - #3830: The message textarea blocks undo of the pasted text
 - #3863: Shift `unescapeHTML` helper function (used to set `isOnlyEmoji` on a message) to converse-headless
+- #3916: Add support for XEP-0461 Message Replies, allowing users to reply to specific messages
+- Don't render unfurls for retracted messages.
 
 ## 12.0.0 (2025-08-28)